add src32 base64, fix bug

+++ src/algorithm/md5.cpp
@@ -141,7 +141,7 @@ void CBasicMD5::update(istream& stream){
 
   while (stream.good()){
     stream.read((char*)buffer, 1024); // note that return value of read is unusable.
-    len=stream.gcount();
+    len = (int)stream.gcount();
     update(buffer, len);
   }
